requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.4.0
pydantic==2.5.1
python-dotenv==1.0.0